Initial commit of files to GitHub, version 0.2
Eclipse project. Requires Groovy. Buildable with Eclipse, and use the
HunspellXML.jardesc to create the hunspell-xml.jar. Also buildable with
Gant. Results in a .jar that can be used in Java, with the addition of
jing.jar from lib/, and one of the groovy-all.jar files from Groovy (see
groovy.codehaus.org).