v0.22 error checking for flags, fixed shortFormatRoute
HunspellXMLUtils.hunspellFlagsToList(...) could throw an exception if
the flag string for a "long" flagType wasn't the right length (multiple
of 2). Trap for this exception and log an error.

HunspellXMLFlagChecker.shortFormatRoute was being used as the key for
checkMap, but the word was being inserted as part of the key when it
shouldn't have been. shortFormatRoute now does not add the word as part
of the key by default - an optional boolean argument of true will cause
the word to be included.