Updated path tracing code for cross product
A flag list that specifies both prefixes and suffixes with cross="true"
set will allow the cross-product of all of those prefixes and suffixes,
as well as single prefixes and affixes. This is now accounted for in the
path-tracing and path-printing algorithms. There's still work that needs
to be done on making circumfix path-tracing work properly.