CLI exits with error on test failure
Modified HunspellXMLCLI to System.exit with an error code when the
Hunspell tests are run and encouter an error while checking the good or
bad word lists.

In order to do this, modified HunspellXMLExporter's options to include
goodPassed, badPassed, goodResults, and badResults.

Modified HunspellXMLTester to record the pass/fail and result list in
HunspellXMLExporter's options map.