Better logging of errors in .dic file
Added better logging for errors in .dic file. Check the words as they
are read in to be able to print out a line number the error occurred on
as well as the complete line. Flag checking methods can be calle with
printError=false param in order to disable printing out errors in places
where useful information like line number cannot be accessed.