Allow ungrouped, unsorted dict word output
Create more options for outputting dictionary words to HunspellXML
format from Hunspell. The words can be output in Hunspell format or in
<w>word</w> format, and also can be grouped by their flag and morph
data, or not grouped at all.

preferWallOfText:true (applies Hunspell and MyThes text format to
dictionary word and thesaurus output)

groupWordByData:false (doesn't group words at all by their flag and
morph data)

On the HunspellXML to Hunspell side, an option can be set to not sort
the dictionary words when they are output: sortDictionaryData (normally
true)

All this allowed the Hunspell test for forbiddenword.dic to pass. Was
that worth it?